deleting items in cart works
