+++ client/components/cart.js
@@ -51,7 +51,7 @@ class Cart extends React.Component {
 
 const mapStateToProps = state => {
   return {
-    cart: state.cart.activeCart
+    cart: state.cart
   }
 }
