+++ client/components/one-product.js
@@ -17,6 +17,7 @@ class OneProduct extends React.Component {
 
   handleAddToCart() {
     const productId = this.props.product.id
+    console.log(this.props.order)
     const orderId = this.props.order.id
 
     this.props.addItem({productId, orderId})
@@ -53,7 +54,7 @@ class OneProduct extends React.Component {
 const mapStateToProps = state => {
   return {
     product: state.product.selectedProduct,
-    order: state.cart.activeCart,
+    order: state.cart,
     user: state.user
   }
 }
