+++ client/store/cart.js
@@ -5,22 +5,17 @@ import axios from 'axios'
  * ACTION TYPES
  */
 const GET_CART = 'GET_CART'
-const GET_HISTORY = 'GET_HISTORY'
 const REMOVE_ITEM = 'REMOVE_ITEM'
 
 /**
  * INITIAL STATE
  */
-const cartState = {
-  activeCart: {},
-  orderHistory: []
-}
+const cartState = {}
 
 /**
  * ACTION CREATORS
  */
 const getCartItems = activeCart => ({type: GET_CART, activeCart})
-const getHistory = history => ({type: GET_HISTORY, history})
 const removeItem = id => ({type: REMOVE_ITEM, id})
 
 /**
@@ -52,34 +47,22 @@ export const addCartItem = ids => {
   }
 }
 
-export const fetchCartHistory = () => {
-  return async dispatch => {
-    try {
-      const {data} = await axios.get('/api/cart/history')
-      dispatch(getHistory(data))
-    } catch (error) {
-      console.error(error)
-    }
-  }
-}
-
 export const deleteCartItem = productIdToRemove => {
   return async dispatch => {
     try {
       await axios.delete(`/api/cart/${productIdToRemove}`)
 
-      dispatch(removeItem(productIdToRemove))
-      //dispatch(fetchCartItems())
+      dispatch(fetchCartItems())
     } catch (error) {
       console.error(error)
     }
   }
 }
 
-export const updateItemQuantity = productId => {
+export const updateItemQuantity = (productId, quantity) => {
   return async dispatch => {
     try {
-      await axios.put(`/api/cart/${productId}`)
+      await axios.put(`/api/cart/${productId}`, quantity)
 
       dispatch(fetchCartItems())
     } catch (error) {
@@ -94,20 +77,7 @@ export const updateItemQuantity = productId => {
 export default function(state = cartState, action) {
   switch (action.type) {
     case GET_CART:
-      return {...state, activeCart: action.activeCart}
-    case GET_HISTORY:
-      return {...state, orderHistory: action.history}
-    case REMOVE_ITEM:
-      const newCart = state.activeCart.activeCart.filter(
-        item => item.id !== action.id
-      )
-      return {
-        ...state,
-        activeCart: {
-          ...state.activeCart,
-          activeCart: newCart
-        }
-      }
+      return action.activeCart
     default:
       return state
   }
