+++ server/db/models/category.js
@@ -1,0 +1,14 @@
+const Sequelize = require('sequelize')
+const db = require('../db')
+
+const Category = db.define('category', {
+  name: {
+    type: Sequelize.STRING,
+    allowNull: false,
+    validate: {
+      notEmpty: true
+    }
+  }
+})
+
+module.exports = Category
